+++ examples/cycle_rainbow.py
@@ -9,18 +9,16 @@ from delfick_project.logging import setup_logging
 import argparse
 import logging
 
-collector = library_setup()
-
-lan_target = collector.configuration["target_register"].resolve("lan")
 
 log = logging.getLogger("cycle_rainbow")
 
 
-async def doit():
-    setup_logging()
+async def doit(collector):
+    lan_target = collector.configuration["target_register"].resolve("lan")
 
     parser = argparse.ArgumentParser()
     parser.add_argument("--reference", required=True)
+    parser.add_argument("--brightness", type=float, default=1)
     args = parser.parse_args()
 
     if args.reference == "_":
@@ -34,7 +32,14 @@ async def doit():
         spread = 1
         color_names = ["blue", "red", "orange", "yellow", "cyan", "green", "blue", "purple", "pink"]
         color_msgs = [
-            Parser.color_to_msg(name, overrides={"res_required": False, "duration": spread})
+            Parser.color_to_msg(
+                name,
+                overrides={
+                    "res_required": False,
+                    "duration": spread,
+                    "brightness": args.brightness,
+                },
+            )
             for name in color_names
         ]
         colors = Pipeline(*color_msgs, spread=spread, synchronized=True)
@@ -51,5 +56,7 @@ async def doit():
         )
 
 
-loop = collector.configuration["photons_app"].loop
-loop.run_until_complete(doit())
+if __name__ == "__main__":
+    setup_logging()
+    collector = library_setup()
+    collector.run_coro_as_main(doit(collector))