+++ examples/toggle_power.py
@@ -3,14 +3,15 @@ from photons_app.special import FoundSerials
 
 from photons_control.transform import PowerToggle
 
-collector = library_setup()
-
-lan_target = collector.configuration["target_register"].resolve("lan")
+from delfick_project.logging import setup_logging
 
 
-async def doit():
+async def doit(collector):
+    lan_target = collector.configuration["target_register"].resolve("lan")
     await lan_target.script(PowerToggle()).run_with_all(FoundSerials())
 
 
-loop = collector.configuration["photons_app"].loop
-loop.run_until_complete(doit())
+if __name__ == "__main__":
+    setup_logging()
+    collector = library_setup()
+    collector.run_coro_as_main(doit(collector))